+++ private/plugins/forum/language/french_canada_utf-8.php
@@ -948,7 +948,12 @@ $LANG_confignames['forum'] = array(
     'smilies_disabled' => 'Réglage par Défaut pour Smileys Box Handicapés',
     'urlparse_disabled' => 'Réglage par Défaut pour l`URL Parse Box Handicapés',
     'use_sfs' => 'Activer Contrôles Arrêter Forum Spam pour de Nouveaux Messages',
-    'allowed_html' => 'Permis HTML'
+    'allowed_html' => 'Permis HTML',
+    'geshi_line_numbers' => 'GeSHi Use Line Numbers?',
+    'geshi_overall_style' => 'GeSHi Overall Code Block Style',
+    'geshi_line_style' => 'GeSHi Line Number Style',
+    'geshi_code_style' => 'GeSHi Code Block Style',
+    'geshi_header_style' => 'GeSHi Header Block Style'
 );
 
 $LANG_configsubgroups['forum'] = array(